remote manager split.
