+++ server/internal/remote/manager.go
@@ -2,11 +2,9 @@ package remote
 
 import (
 	"fmt"
-	"os/exec"
 	"time"
 
 	"m1k1o/neko/internal/gst"
-	"m1k1o/neko/internal/remote/clipboard"
 	"m1k1o/neko/internal/remote/xorg"
 	"m1k1o/neko/internal/types"
 	"m1k1o/neko/internal/types/config"
@@ -207,56 +205,4 @@ func (manager *RemoteManager) ChangeResolution(width int, height int, rate int) 
 	}
 
 	return nil
-}
-
-func (manager *RemoteManager) Move(x, y int) {
-	xorg.Move(x, y)
-}
-
-func (manager *RemoteManager) Scroll(x, y int) {
-	xorg.Scroll(x, y)
-}
-
-func (manager *RemoteManager) ButtonDown(code int) error {
-	return xorg.ButtonDown(code)
-}
-
-func (manager *RemoteManager) KeyDown(code uint64) error {
-	return xorg.KeyDown(code)
-}
-
-func (manager *RemoteManager) ButtonUp(code int) error {
-	return xorg.ButtonUp(code)
-}
-
-func (manager *RemoteManager) KeyUp(code uint64) error {
-	return xorg.KeyUp(code)
-}
-
-func (manager *RemoteManager) ReadClipboard() string {
-	return clipboard.Read()
-}
-
-func (manager *RemoteManager) WriteClipboard(data string) {
-	clipboard.Write(data)
-}
-
-func (manager *RemoteManager) ResetKeys() {
-	xorg.ResetKeys()
-}
-
-func (manager *RemoteManager) ScreenConfigurations() map[int]types.ScreenConfiguration {
-	return xorg.ScreenConfigurations
-}
-
-func (manager *RemoteManager) GetScreenSize() *types.ScreenSize {
-	return xorg.GetScreenSize()
-}
-
-func (manager *RemoteManager) SetKeyboardLayout(layout string) {
-	_ = exec.Command("setxkbmap", layout).Run()
-}
-
-func (manager *RemoteManager) SetKeyboardModifiers(NumLock int, CapsLock int, ScrollLock int) {
-	xorg.SetKeyboardModifiers(NumLock, CapsLock, ScrollLock)
 }
